fix potential bug

+++ ft_printf.h
@@ -18,13 +18,13 @@
 # include <unistd.h>
 # include <stdint.h>
 
-int		ft_printf(const char *str, ...);
-void	ft_bzero(void *s, size_t n);
-void	*ft_calloc(size_t nmemb, size_t size);
-char	*ft_itoa(long long int n, char *base);
-char	*ft_uitoa(unsigned long long n, char *base);
-int		ft_putchar_fd(char c, int fd);
-int		ft_putstr_fd(char *s, int fd);
-int		ft_putstr_nil_fd(char *s, int fd);
-size_t	ft_strlen(const char *str);
+int			ft_printf(const char *str, ...);
+void		ft_bzero(void *s, size_t n);
+void		*ft_calloc(size_t nmemb, size_t size);
+char		*ft_itoa(long long int n, char *base);
+char		*ft_uitoa(unsigned long long n, char *base);
+ssize_t		ft_putchar_fd(char c, int fd);
+ssize_t		ft_putstr_fd(char *s, int fd);
+ssize_t		ft_putstr_nil_fd(char *s, int fd);
+size_t		ft_strlen(const char *str);
 #endif
